requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
supabase>=2.0.0

# AI Provider SDKs
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0

# Async Support
aiohttp>=3.9.0
asyncio>=3.4.3

# Database and Queue Management
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Image Processing
Pillow>=10.0.0
requests>=2.31.0

# Scheduling and Background Tasks
schedule>=1.2.0
celery>=5.3.0  # Optional: for distributed task queue

# Alerting
sendgrid>=6.10.0  # For email alerts
slack-sdk>=3.24.0  # Optional: for Slack notifications

# Monitoring and Logging
structlog>=24.0.0
prometheus-client>=0.19.0

# Utilities
pydantic>=2.5.0  # For data validation
tenacity>=8.2.0  # For retry logic
cachetools>=5.3.0  # For in-memory caching

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6  # For file uploads
websockets>=12.0
jinja2>=3.1.3  # For HTML templates

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0

# Optional Performance Enhancements
redis>=5.0.0  # For distributed caching
uvloop>=0.19.0  # Faster async event loop